Matrix

+++ day20/Solution.hpp
@@ -10,7 +10,7 @@ std::vector<int> parse();
 struct Solution
 {
     int solve(std::vector<int> );
-    long solve_part2(std::vector<int> );
+    long long solve_part2(std::vector<int> );
 };
 
 }
